feat(react): refactor util `getModuleFederationConfig` to avoid to pass function to determinate the remote url
