+++ packages/react/src/module-federation/with-module-federation-ssr.ts
@@ -1,24 +1,5 @@
 import { ModuleFederationConfig } from '@nx/devkit';
-import { readCachedProjectConfiguration } from 'nx/src/project-graph/project-graph';
 import { getModuleFederationConfig } from './utils';
-
-function determineRemoteUrl(remote: string) {
-  const remoteConfiguration = readCachedProjectConfiguration(remote);
-  const serveTarget = remoteConfiguration?.targets?.['serve-server'];
-
-  if (!serveTarget) {
-    throw new Error(
-      `Cannot automatically determine URL of remote (${remote}). Looked for property "host" in the project's "serve-server" target.\n
-      You can also use the tuple syntax in your webpack config to configure your remotes. e.g. \`remotes: [['remote1', 'http://localhost:4201']]\``
-    );
-  }
-
-  const host = serveTarget.options?.host ?? 'http://localhost';
-  const port = serveTarget.options?.port ?? 4201;
-  return `${
-    host.endsWith('/') ? host.slice(0, -1) : host
-  }:${port}/server/remoteEntry.js`;
-}
 
 export async function withModuleFederationForSSR(
   options: ModuleFederationConfig
@@ -26,7 +7,7 @@ export async function withModuleFederationForSSR(
   const reactWebpackConfig = require('../../plugins/webpack');
 
   const { sharedLibraries, sharedDependencies, mappedRemotes } =
-    await getModuleFederationConfig(options, determineRemoteUrl, {
+    await getModuleFederationConfig(options, {
       isServer: true,
     });
