+++ packages/react/src/module-federation/with-module-federation.ts
@@ -1,26 +1,7 @@
 import { ModuleFederationConfig } from '@nx/devkit';
-import { readCachedProjectConfiguration } from 'nx/src/project-graph/project-graph';
 import { getModuleFederationConfig } from './utils';
 import ModuleFederationPlugin = require('webpack/lib/container/ModuleFederationPlugin');
 import type { AsyncNxWebpackPlugin } from '@nx/webpack';
-
-function determineRemoteUrl(remote: string) {
-  const remoteConfiguration = readCachedProjectConfiguration(remote);
-  const serveTarget = remoteConfiguration?.targets?.serve;
-
-  if (!serveTarget) {
-    throw new Error(
-      `Cannot automatically determine URL of remote (${remote}). Looked for property "host" in the project's "serve" target.\n
-      You can also use the tuple syntax in your webpack config to configure your remotes. e.g. \`remotes: [['remote1', 'http://localhost:4201']]\``
-    );
-  }
-
-  const host = serveTarget.options?.host ?? 'http://localhost';
-  const port = serveTarget.options?.port ?? 4201;
-  return `${
-    host.endsWith('/') ? host.slice(0, -1) : host
-  }:${port}/remoteEntry.js`;
-}
 
 /**
  * @param {ModuleFederationConfig} options
@@ -32,7 +13,7 @@ export async function withModuleFederation(
   const reactWebpackConfig = require('../../plugins/webpack');
 
   const { sharedDependencies, sharedLibraries, mappedRemotes } =
-    await getModuleFederationConfig(options, determineRemoteUrl);
+    await getModuleFederationConfig(options);
 
   return (config, ctx) => {
     config = reactWebpackConfig(config, ctx);
